fix: removes Tailwind artifacts from ST frontend

+++ src/backend/src/router/study_tracker/dtos/input_dtos.py
@@ -56,5 +56,5 @@ class CreateCurricularUnitInputDto(BaseModel):
     name: str
     
 class CreateGradeInputDto(BaseModel):
-    value: int
-    weight: int
+    value: float
+    weight: float
